+++ extensions/homeassistant/src/components/battery/list.tsx
@@ -1,7 +1,8 @@
 import { List, showToast, Toast } from "@raycast/api";
 import { useState } from "react";
-import { useHAStates } from "../hooks";
-import { StateListItem, useStateSearch } from "./states";
+import { useHAStates } from "../../hooks";
+import { StateListItem, useStateSearch } from "../states";
+import { sortBatteries } from "./utils";
 
 export function BatteryList(): JSX.Element {
   const [searchText, setSearchText] = useState<string>();
@@ -20,14 +21,7 @@ export function BatteryList(): JSX.Element {
     return <List isLoading={true} searchBarPlaceholder="Loading" />;
   }
 
-  const sortedStates = states?.sort((s1, s2) => {
-    const s1v = parseFloat(s1.state);
-    const s2v = parseFloat(s2.state);
-    if (s1v === s2v) {
-      return 0;
-    }
-    return s1v < s2v ? -1 : 1;
-  });
+  const sortedStates = sortBatteries(states);
   return (
     <List searchBarPlaceholder="Filter by name or ID..." isLoading={isLoading} onSearchTextChange={setSearchText}>
       {sortedStates?.map((state) => <StateListItem key={state.entity_id} state={state} />)}