+++ extensions/homeassistant/src/components/update/utils.ts
@@ -1,0 +1,6 @@
+export interface HACSRepo {
+  name: string | undefined;
+  display_name: string | undefined;
+  installed_version: string | undefined;
+  available_version: string | undefined;
+}