+++ extensions/homeassistant/CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ## [Menu] - 2023-07-25
 
-- Add Mediaplayer Menu Bar command
+- Add Media Player Menu Bar command
 - Add Entities Menu Bar command to allow to add specific entities in the Menu Bar
 - Add Lights Menu Bar command
 - Add Covers Menu Bar command
+- Notifications menu now contains battery low battery states as well as updates
 
 ## [Modernize] - 2023-07-16